refactor: extract StandingsRowView from StandingsView (Phase 2)
