+++ ibl5/tests/PHPStanRules/Fixtures/classes/Standings/StandingsRowView.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+$unsafeVariable = 'test';
+
+?>
+<html>
+<body>
+    <?= $unsafeVariable ?>
+</body>
+</html>
